// src/app/models/sales/salesOrderReference.ts
export class SalesOrderReference {
    salesOrderId: number;
    salesOrderNumber: string;
    itemMasterId: number;
    salesOrderPartId: number;
    customerId: number;
    customerName: string;
    partNumber: string;
    partDescription: string;
    quantity: number;
    managementStructureId: number;
    conditionId: number;
}